+++ src/Engine/CMakeLists.txt
@@ -47,7 +47,6 @@ set(SRCS
         Util/Json.cpp
         Util/Profiling.cpp
         Util/Settings.cpp
-        Texture/Texture2D.cpp
         Texture/TextureAsset.cpp
         Script/ScriptFile.cpp
         Input/Input.cpp
@@ -95,7 +94,6 @@ set(HEADERS
         Util/Json.hpp
         Util/Profiling.hpp
         Util/Settings.hpp
-        Texture/Texture2D.hpp
         Texture/TextureAsset.hpp
         Script/ScriptFile.hpp
         Input/Input.hpp
